Core: small fixes in domain entities and added migration Initial_create

+++ TestingSystem/Core/Domain/Entites/Group.cs
@@ -6,7 +6,7 @@ namespace Core.Domain.Entites
     {
         public short Id { get; private set; }
         public string Title { get; private set; }
-        public short ExaminerId { get; private set; }
+        public int ExaminerId { get; private set; }
         public Examiner Examiner { get; private set; }
         public short TestId { get; private set; }
         public Test Test { get; private set; }
@@ -17,7 +17,7 @@ namespace Core.Domain.Entites
             StudentGroups = new List<StudentGroup>();
         }
 
-        public Group(string title, short examinerId, short testId)
+        public Group(string title, int examinerId, short testId)
         {
             Title = title;
             ExaminerId = examinerId;
